migrate to agent_common
